Remove uneeded chrome.KeepState()
Starting chrome with chrome.KeepState() was a relic from when we set up
consent directly before Chrome started instead of via a call to Chrome.
This was actually causing a lot of false positives which make it hard to
debug other consent issues. (If the machine already had consent turned
on from a different test, it made it look like our consent setting was
working.)

Note that these changes will cause some breakpad real-consent tests to
start failing much more often until a version of Chrome with
https://chromium-review.googlesource.com/c/chromium/src/+/2145676 is
present, but since those tests aren't critical anyways, I'm checking
this in.

BUG=chromium:984807
TEST=tast run ... ui.ChromeCrashReporterMetrics*ui.ChromeCrashLoggedIn.* ui.ChromeCrashNotLoggedIn.*

Change-Id: I4794ef95c3792a99c0a8c3016084227a0d7dbfd1