Remove uneeded chrome.KeepState()
Starting chrome with chrome.KeepState() was a relic from when we set up
consent directly before Chrome started instead of via a call to Chrome.
This was actually causing a lot of false positives which make it hard to
debug other consent issues. (If the machine already had consent turned
on from a different test, it made it look like our consent setting was
working.)

BUG=chromium:984807
TEST=tast run ... ui.ChromeCrashReporterMetrics*ui.ChromeCrashLoggedIn.* ui.ChromeCrashNotLoggedIn.*

Change-Id: I4794ef95c3792a99c0a8c3016084227a0d7dbfd1