Create new test: firmware.Config
Previously, we were testing that we could correctly load
fw-testing-configs as a side-effect of firmware.CheckBootMode. But it
makes more sense to move this into its own test. This way we can verify
that each board is able to load its configs.

Generalized config functionality, such as inheritance, would be better
handled in a config_test.go.

BUG=b:151832522
TEST=firmware.Config
TEST=firmware.CheckBootMode

Change-Id: Ic31622a731c884621a52f9aead9cbb3b8781da44