Create new test: firmware.Config
Previously, we were testing firmware configs as a side-effect of
firmware.CheckBootMode. But as we start to develop more nuanced Config
functionality, such as inheriting from parent/grandparent configs and
from DEFAULTS, and such as overriding with model-specific configs, it
will be useful to have a separate test specifically for Config.

BUG=b:151832522
TEST=firmware.Config
TEST=firmware.CheckBootMode

Change-Id: Ic31622a731c884621a52f9aead9cbb3b8781da44