ChromeCrashLoop: stop test after crash loop found
Once Chrome crashes in crash-loop mode, the testuser is normally logged
out. This causes session_manager to quit its run loop and restart. The
new, restarted session_manager doesn't have our test args. In
particular, it doesn't have the mock consent arguments we use for
breakpad, so if the test's kill-and-look-for-crash-files loop executes
another time, it will fail because we no longer have mock consent.

BUG=chromium:994987
TEST=tast run ... ui.ChromeCrashLoop*mock_consent

Change-Id: Ida8d0f4c723f92abe3ad674df8fa4b4c4ace554f