Update ui.TabSwitchCUJ scenario slightly
As I quickly checked the behavior of ui.TabSwitchCUJ test case,
I've noticed that the subtest of 'Google News' does not work
as expected -- somehow opened articles stay empty and not get
redirected, and the tab-switching is much faster than actual.
Somehow WPR does not properly handle that case, though it is
not clear why it fails.

This CL removes the case of 'Google News', and introduces a new
'Reddit' case instead. It also updates the wprgo data.

BUG=None
TEST=tast run ui.TabSwitchCUJ

Change-Id: Id8e917e8330a6b58e0b9024d2039e45f3dd07a87