dev.RemoteDesktop: Limit the number of sessions by hwdeps
This is a speculative fix to limit the number of sessions. It seems that
the test account is throttled by the CRD backend, so the test is failing
with a periodic pattern. The model list is handcrafted to cover various
platforms.

Recent test results on stainless with a periodic failing pattern:
https://stainless.corp.google.com/search?view=matrix&row=build&col=test&first_date=2020-03-01&last_date=2020-04-16&master_builder_name=%5Emaster-release%24&branch=%5Emaster%24&test=dev.RemoteDesktop&exclude_cts=false&exclude_not_run=true&exclude_non_release=true&exclude_au=true&exclude_acts=true&exclude_retried=true&exclude_non_production=false

BUG=b:134038021,b:151111783
TEST=Pass dev.RemoteDesktop and dev.RemoteDesktop.test on nocturne.

Change-Id: I0c240bb6c3234ebee7784c22045bdb41d41814f2