Update patch set 3
Patch Set 3:

(13 comments)

> Patch Set 3:
> 
> What do we expect to be recorded here? If we are just recording environment noise, won't it be flaky?
> Instead of quality check, can we change this to something like CaptureVolume check or gain check?

We thought the devices in the lab should not be flaky due to the environment noise.
The problem we want to catch is the static noise, which should not happen in the environment.
We can deploy it first to check whether it works. (And modify the threshold if needed)

Patch-set: 3