tast-tests: add platform.CrosHardwareVerifier remote test
In order to ensure no regression of hardware_verifier occurs in future
builds, this CL creates a tast test for check the integrity and
consistency of the report generated by hardware_verifier.  Since the
generation of report needs a reboot, we choose remote test for this
case.

BUG=b:150830664
TEST=tast run $DUT platform.CrosHardwareVerifier

Change-Id: I5415d06c699743306aaeecc8873f26905ee47180
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2152060
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>