arc.ContainerMount: Update with temporary /home/root bind-mount.
As part of go/cros-non-root-mount-ns, we're going to need to
temporarily bind-mount /home/root to itself, to allow for a smoother
migration of system services that are not using daemon-store.

Update the test to reflect this bind-mount, which manifests as /data
being a slave mount inside the container (since /data comes from inside
/home/root).

(See crrev.com/c/2150054 for details).

BUG=chromium:1069501
TEST=Passes with crrev.com/c/2150054.

Change-Id: I8e7f20167bc01116b72666b2a76def2fd3c03da0