[Tast] Fix flaky hotseat scroll test [part 2]
ui.HotseatScrollPerf is flaky. It may be caused by unstablet UI right
after switch between clamshell and tablet. In this CL, wait for one
second to stabilize UI. If it solves the problem, a following CL should
be landed for the real fixing. In addition, it is not meaningful to
check hotseat scroll in tablet when there is no internal display. So if
there is no internal display, skip the test for tablet mode.

BUG=chromium:1064459
TEST=manually

Change-Id: I6578979814bc92434a5b203134cd39e17559ba45