camera.Connector: Wait camera socket ready before running test
The camera socket is created/destroyed by Chrome and it take some time
to setup the socket properly. The camera connector library would fail to
init() if the socket is not ready yet. Ensure the camera socket is ready
before executing the test binary to avoid race condition.

Note that we don't want to poll this condition in the connector, because
the socket file might be missing as intended in guest OS sandbox due to
the permission setting, and init() should fail immediately in that case.

The checking logic is same as the one we used in cros-camera service
https://chromium.git.corp.google.com/chromiumos/platform2/+/f0f78cfa3b1069c7f33b378699c3d7ede790eff7/camera/hal_adapter/camera_hal_server_impl.cc#97

BUG=b:151047930
TEST=tast run <nocturne> audio.Microphone camera.Connector, where
audio.Microphone is the test scheduled just before camera.Connector.

Change-Id: I1d8ee45c985fa8fd8a5a0e4f36d5d21222b78c15