dutfs: Replace Dial with NewClient.
It is often the case that a remote test wants to use the dutfs package
as well as other gRPC services. Currently the dutfs package only
provides the Dial function which establishes a new gRPC connection to
the DUT, thus such tests always end up establishing multiple gRPC
connections, which is waste of runtime.

This change introduces the NewClient function which returns a client
for an existing gRPC connection, and removes the Dial function.

Also rewrites the crash.ReporterStartup test to use the new function.
Actually it is an example of a test that used to have to establish
two indepdendent gRPC connections if it uses the dutfs package.

BUG=chromium:1071736
TEST=fast_build.sh -T

Change-Id: I4b78708bf1dc45a09c81e2fb1cca654e48c5a296
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2153233
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>