tast-tests: Test a11y focusability of a node.
This CL adds a regression test for b/154173452, by adding a test for a
node with no description.

Tast accessibility test app has two Sliders, but technically there is no
difference between these two. This CL changes one of them to have no
contentDescription so that we can test for a node without any
description.

BUG=b:154173452
TEST=tast run DUT_IP arc.accessibility* (after deploying apk).
Failing without crrev/c/2153450, passing with it.

Change-Id: I2e09574a720da80e9777145f7051fca53bcc8606