tast-tests/webrtc: add video conf call cases to RTCPeerConnectionPerf
This CL adds Go+JS code and cases to support video-call like scenarios,
which are simulated by having a loopback RTCPeerConnection and a number
of videos being played back, all in a NxN grid, see e.g.
- 3x3 https://imgur.com/a/mh3GTS1
- 4x4 https://imgur.com/a/JJsdqIA
etc.

To better replicate Meet, the local loopback is using VP8 whereas the
"remote" videos are using VP9.

The grid is created via modifying the <div> containing the "remote" video
into a CSS display=grid of the appropriate dimension, and adding <video>s
playing the videoURL that is passed as parameter in the new test cases.

BUG=b:154005031
TEST=tast run DUT webrtc.RTCPeerConnectionPerf.*multi* on kohaku

Change-Id: I93e26b12af4d6fa6cc0563c0091702cb6e3f451e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2153510
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>