tast-tests/webrtc: add video conf call cases to RTCPeerConnectionPerf
This CL adds Go+JS code and cases to support video-call like scenarios,
which are simulated by having a loopback RTCPeerConnection and a number
of videos being played back, all in a NxN grid, see e.g.
- 3x3 https://imgur.com/a/mh3GTS1
- 4x4 https://imgur.com/a/JJsdqIA
etc.

The grid is created via modifying the <vid> containing the video into
a container-grid of the appropriate dimension, and adding videoURL
that is passed as parameter in the new test cases.

BUG=b:154005031
TEST=tast run DUT webrtc.RTCPeerConnectionPerf.*multi* on kohaku

Change-Id: I93e26b12af4d6fa6cc0563c0091702cb6e3f451e