crostini: use crostini installer in preconditions
This increases coverage of crostini installer and crostini manager and
makes the test environment more similar to the real environment.

To use the same code paths as user installation we need to have lxd
running in Termina download the container from an LXD server, but we
want to control the image that is used.  Thus we start a local fake
lxd server in the tast process and point lxd to that server.

Design proposal at go/tast-crostini-setup-proposal

BUG=chromium:994040
TEST=Run artifact/download/download buster tests on eve

Change-Id: I3dde41cbb7d14ca929d92d86b0cf6aab84f4b735