Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > hello, zuan@:
> > to put all of us on the same page, could you please elaborate "the way WaitForUserMount() check for partition"?
> > 
> > I will start review until zuan@'s comment comes in. Maybe the decision would be clear then.
> 
> Thanks for bringing this up. Sorry John, I somehow didn't see the partition part in your comment.
> Yeah, I'd also like to know how the mountinfo check could fail to work.
> Here is a brief summary of what I see and understand:
> 
> cryptohome::Platform reads /proc/<pid>/mountinfo to get the list of mount points, whereas the go library basically pulls them from /proc/self/mountinfo. There are some backward compatibility implementations in the go lib too, but beside that, the logic is identical.
> When calling the cryptohome client from Tast to read mountinfo (b/153516375#comment4), <pid> and self should be identical. Also, IIUC, the tests are run with root privileges, so it should be able to see all mount points? Not sure why the go lib sees a different list...


Your understanding is generally correct.

I'm a little bit inclined to propose that we parse /proc/self/mountinfo ourselves to get to the bottom of this. Chrome OS runs a limited selection of Linux Kernel version, so we don't have that much to support (no backward compatibility stuff). Meanwhile, if we want to try getting the mountinfo of cryptohome to see if it's mount namespace messing with this, we can (if we parse it ourselves), and that's not supported by the lib.

Patch-set: 3
Reviewer: Gerrit User 1314760 <1314760@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>