network: separate SGI capabilities
The SGI capability of different channel widths are different bits in
capability bitmap. Let's separate them and only check the ones of HT20
and HT40 in network.WifiCaps as monroe does not yet support 802.11ac.

This cl will fix the failure of network.WifiCaps on monroe.

For 802.11ac checks, another sub-test with HWDep will be added later.

BUG=b:154552359
TEST=`tast run $DUT network.WifiCaps` on monroe DUT

Change-Id: I6d965cabcb833739dd59492d84f6662be8808a04