Crostini: Migrate to hwdeps
Start using hwdeps instead of softwareDeps for defining the set of
boards that run crostini on the CQ.

Only change a single test to check that everything works before
migrating the rest of the tests.

Unfortunately there is no "OR" combinator for hwdeps so we can only
choose a blacklist for platform or model.  Choosing model so that we can
enable more models that are stable without the whole platform being
stable. The previous softwareDeps version was using platform
blacklisting.

I grabbed the list of models that should be unstable from this query:
https://stainless-dev.googleplex.com/search?view=matrix&row=board_model&col=build&first_date=2020-04-15&last_date=2020-04-21&test=crostini.Sanity.artifact_unstable&exclude_cts=false&exclude_not_run=false&exclude_non_release=true&exclude_au=true&exclude_acts=true&exclude_retried=false&exclude_non_production=false

BUG=None
TEST=Ran stable and unstable test on stable and unstable board and
checked that it was run/skipped correctly.

Change-Id: I311ed9f9da3bac939f2cb25753b92881178ccc89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2158645
Tested-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>