network: ping: Handle the case when all packets are lost.
When all packets are lost, ping will return non zero value and be
regarded as command failure. However, we sometimes want to verify
loss of connection and will have to identify 100% loss but not command
error. Update the logic so it can return valid result in this case.

The behavior will be used in later WiFi tests like network_WiFi_CSA.

BUG=b:154879577
TEST=fast_build.sh -T
TEST=run network.VPNConnect.l2tp_ipsec_psk test in crrev.com/c/2117039

Change-Id: I18b9e54aeee114904de77c023c44bece6ede79a2