tast-tests/webrtc: set landscape mode if tablet for MediaRecorderAccelerator
Tablet devices such as e.g. scarlet capture from real webcams
responding to the orientation of the device, so if the device is
e.g. in portrait mode, so will the capture be. The captured size
(720x1280) is then too large for the hw encoder (max: 1920x1088)
and the test fails.

This CL forces landscape mode for MediaRecorderAccelerator.vp8_cam
test case and if the DUT is in tablet mode.

BUG=chromium:1071979
TEST=tast run DUT webrtc.MediaRecorderAccelerator.vp8_cam
  on scarlet and kohaku

Change-Id: I44a7a48f6022769707dc59296331af7b5901b033