Limit the app installation of ui.TaskSwitchCUJ'*'
The test log in stainless revealed that most of the failures
happen during the app installation. The network is limited in the
lab and sometimes it fails to install which leads to the failure
of the entire scenario.

With this CL, test installs only one app -- gmail -- and uses
others through web. This CL also keeps the play-store app window
too (IIRC I wrote the closing logic since it didn't work well with
the test scenario using some test APIs -- now it works well).

BUG=chromium:1065519
TEST=tast run ui.TaskSwitchCUJ'*'

Change-Id: I8665c0365e3eb65ae6ed1a1738c08b86536b40aa