tast: Add arc.MotionInput test
This CL adds a test for input, which focuses on the location of
MotionEvents.

MotionEvents include touch, mouse, stylus, and other events that are not
KeyEvents. Motion events are often targeted to specific coordinates,
especially in the case for mouse and touchscreen.

This test installs and runs an Android application, injects various
input events into ChromeOS through uinput, and verifies that those events
were received by the Android application in the correct screen location.

Since the location of MotionEvents in Andorid depends on several WM
features (e.g. visibility of the caption, tablet/clamshell mode, etc.),
this test runs several sub-tests that verify MotionEvent functionality
in various WM modes.

BUG=b:152436345
TEST=tast run DUT arc.MotionInput.vm

Change-Id: Idf20a3bb3e39f4381e1d2476b0e7e742192f3832