Update patch set 12
Patch Set 12: Verified+1 Commit-Queue+1

(1 comment)

Hi reviewers, I have updated the CL.
Please have a look when you have time.

After the discussion with tfiga@ at b/153387538, it is determined that we build the test executables on the Android side at least for the time being.
I have uploaded a CL on the Android side (ag/10930995) which contains the source code, the build rule, and an explanation on how to update the executables in GCS.

Currently we do not refer to the source code from the Tast test side.
Actually, it seems that other Tast tests using APKs built on the Android side do not do that, either.
I'd like to know if there is an appropriate way to refer to it from the test side.

Thanks!

Patch-set: 12
Reviewer: Gerrit User 1394308 <1394308@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Label: Verified=+1