DualShock 3 test
The gamepad.DS3 test checks that the mappings on the dualshock 3 are
correct. The uhid library is used to create a virtual dualShock 3,
replay a hid recording and record the received events via a javascript
script. These are then compared with an array of expected button
presses.

BUG=chromium:1081851

TEST=tast -verbose run <DUT_IP> gamepad.DS3

Change-Id: I45a99ca3c97c1de8f18cb2d4d1ae1480b1f7d42d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2173302
Tested-by: José Torreguitar <jtguitar@google.com>
Commit-Queue: José Torreguitar <jtguitar@google.com>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Auto-Submit: José Torreguitar <jtguitar@google.com>