Update patch set 5
Patch Set 5: Code-Review+1

I'm not really sure what the practical effects of promoting this to critical are on the grunt-kernelnext development. 

I'm fine with this in theory but don't want to block people's work. I'll defer to the tast owners on the practical realities.

Random thought: should we have a non-grunt-kernelnext subtest (critical, doesn't run on grunt-kernelnext) and a grunt-kernelnext subtest (only runs on grunt-kernelnext, informational)? Or is that getting too complex?

Patch-set: 5
Reviewer: Gerrit User 1300299 <1300299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1