tast-test: hwsec: AttestationNoExternalServer
This CL creates a duplicate of hwsec.Attestation, but disabled. The
purpose of this new test item is to deploy the external-server-free test
infrastucture for attestation flow and put the attestation flow terting
in CQ.

The local testing infrastructure will be adopted by this test item in
the near future.

BUG=b:154672520
TEST=hwsec.AttestationNoExternalServer

Change-Id: I01e5dbb87da38ad068b136d5a3476b96e5d3017b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2172559
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>