platform.P2PServer: Avoid entering infinite loop.
Without checking the context cancellation, queryP2PServices does not
return an error even if the context deadline is already reached or
it is canceled during the call. This will cause waitP2PService to
enter an infinite loop on test timeout.

This change does not fix test failures, but makes the errors
easier to understand.

BUG=b:153182041
TEST=tast run DUT platform.P2PServer

Change-Id: Ief01d13c5cd91b3521386c642158c76eb2233362