wificell: router: Trim logs before NewRouter.
Only collect the logs during the lifetime of the Router object to
reduce log size. Also, export CollectLogs function as we're going to
introduce precondition for TestFixture, the Router will only be closed
once. We'll have to move it out so that each test can collect their own
log in its lifecycle.

NOTE: The path of collected logs is changed in this cl.

BUG=b:154662380
TEST=`tast run $DUT wifi.SimpleConnect\*` and check if the logs is
trimmed.

Change-Id: I204c8aa9967b3bbc206c2e5f9480b23cf1a2687d