[tast] Fix constant failure of ui.AssistantTimeQuery.
Previously we used the html fallback message to parse the time result.
As the fallback text has been deprecated now, it no longer contains any
meaningful information, which causes the test to fail. Now in this CL,
we instead use the html string, which is also a more stable source.

BUG=b:153914355
TEST=run all related tast tests.

Change-Id: If24363b62d98b0f22e3b37108ae054298ed7dfc5