Update patch set 11
Patch Set 11:

> Patch Set 10:
> 
> PTAL. Now cleanup is performed repeatedly at the beginning of the test to make sure previous tests don't affect this one.

For cleanup, I'm wondering if it makes sense to add a cryptohome action to remove an account given a shadow cryptohome path (eg. RemoveByVaultPath) for testing purposes? The current setup is a bit fragile and if we have a specific action that does that, all we need to do in the test is iterate through existing accounts and cleanup the existing users.

Patch-set: 11
Reviewer: Gerrit User 1241666 <1241666@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>