crostini: change tests to hwdeps for stable boards
Using hwdeps for stable boards allows us to change the list of boards in
a single CL instead of 2+ cls with the current swdeps approach.

The hwdeps approach has been working well for the sanity test over the
weekend, switching the rest of the tests over now.

Note that there is another cl out to switch the graphics trace replay
test to hwdeps as well.  I can't delete the swdeps code until that cl
lands as well.

BUG=None
TEST=Verified AudioSanity test skipped/not skipped on eve

Change-Id: I1b04d1c37685864d43667c52584eb09597667772
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2166369
Tested-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>