crostini: change tests to hwdeps for stable boards
Using hwdeps for stable boards allows us to change the list of boards in
a single CL instead of 2+ cls with the current swdeps approach.

The hwdeps approach has been working well for the sanity test over the
weekend, switching the rest of the tests over now.

Note that there is another cl out to swith the graphics trace replay
test to hwdeps as well.  I can't delete the swdeps code until that cl
lands as well.

BUG=None
TEST=Verified AudioSanity test skipped/not skipped on eve

Change-Id: I1b04d1c37685864d43667c52584eb09597667772