tast-tests: Port cheets_DragDropTest to Tast
This CL ports the autotest cheets_DragDropTest to Tast. It tests drag
and drop operation from Chrome to ARC.

Change from the original test: The original test injects a mouse
event sequence that was captured from a real device, but this Tast test
uses ash.MouseDrag. Other than that, this test tries to imitate the
original approach as much as possible, but further simplification may be
possible.

BUG=b:153402732
TEST=tast run eve-DUT arc.DragDrop

Change-Id: I42029f434a0b77cd7313a4296e4d94b5c33d4b8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2166378
Tested-by: Tetsui Ohkubo <tetsui@chromium.org>
Commit-Queue: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>