tast-tests: hwsec: Add chaps helper function that create all keys that should be covered
Currently many tests need to test different aspects of the same set of keys,
so there are duplicate code that create the same set of keys in each of those
tests.

This change adds the code that creates all those keys with a single function
call.

BUG=b:155088532
TEST=tast run $dut hwsec.Chaps* with CLs later down the chain.

Change-Id: I25302171171e198a634ad06a74466ff5f604d15f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2167954
Tested-by: John L Chen <zuan@chromium.org>
Commit-Queue: John L Chen <zuan@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>