Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Sorry for the delay. I don't think it will break anything. But I am not sure who is going to monitor the perf results? Collecting the data seems like a distraction. More on the issue.
> 
> If you'd rather not collect data I can remove it from the suites.  I do agree the data tends to be noisy with these low resolutions anyway.

As long as you are ok that we just monitor it for pass/fail and accept that we don't actively watch the numbers then we can run it for now. (But somebody should be watching numbers at times.) Now if one day we have to cull tests I rather sample high res (or very popular) ones than every low res, but we don't need to cull now.

Patch-set: 2