Use "crash_selinux_(un)stable" deps.
Per the linked bug, platform.SelinuxViolation is broken on 5.4 kernels.
I believe that the root cause is that on 5.4, selinux audit logs are
only written to /var/log/audit.log, and not to journald. The fix will be
to change anomaly_detector to read .log files instead of the journal,
but this might take time and will need somewhat extensive verification
before we can submit it.

In the meantime, since 5.4 isn't (yet) on any stable boards, we want to
un-mark these tests as critical on these boards to allow lab development
to proceed.

BUG=b:155009503
TEST=tast run $DUT platform.SelinuxViolation.mock_consent

Cq-Depend: chromium:2169022
Change-Id: Ifd40634f0ac39b290bb834d18ba3f6a8313d3213