UncleanShutdown: Stash suspend file.
If the powerd_suspended file happens to exist when the test starts,
crash_collector --boot_collect will see it and assume that the
"unclean shutdown" happened because of a battery failure (and not report
it). Work around this by stashing the suspend file if it exists.

BUG=chromium:1041639
TEST=tast run $DUT platform.UncleanShutdownCollector
TEST=create powerd_suspended file; run test; check that it's still there.

Change-Id: I5acbc5e05178fcf82b9675fbe2cfbfa5eed8c58a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2171928
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Joon Ahn <joonbug@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>