Fix ui.LauncherDragPerf for nocturne/kodama/scarlet
When conbined with ui.HotseatScrollPerf (with the same precondition),
there would be a lot of app icons in the shelf, which makes a lot
of location change events on transitioning to the clamshell mode.
As the result, the drag gesture would fall into a wrong location
without waiting.

BUG=chromium:1076520
TEST=tast run ui.HotseatScrollPerf ui.LauncherDragPerf on nocturne

Change-Id: Id68a415ede85445b01fa875492d6ad20345bb17b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2173223
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>