Update patch set 3
Patch Set 3:

I'm not a huge fan of this timestamp + sed-command stuff. It seems to work I guess (or at least, we don't notice when it doesn't, because these logs are non-critical), but I'm pretty sure it doesn't handle things like log rotation correctly.

It would be extra nice if we could make use of src/chromiumos/tast/local/syslog/reader.go, but that won't work, as it's not factored for remote tests, and we're not running Tast gRPCs on the AP anyway. But is it possible to do something similar in spirit? e.g., a running SSH pipe/buffer that tails the syslog, handles EOF, etc., and CollectLogs() just pulls from that buffer?

Let me know if you think that is feasible (or, not worthwhile for some reason) and I can come back to review this one as-is instead.

Patch-set: 3
Reviewer: Gerrit User 1132414 <1132414@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>