Update patch set 6
Patch Set 6:

(1 comment)

> Patch Set 3:
> 
> > Let me know if you think that is feasible (or, not worthwhile for some reason) and I can come back to review this one as-is instead.
> 
> Thanks for the suggestion. I think it is feasible, let me try to update my cl.

Hmm... after some trials, spawning a "tail" and collect its output is possible. But with Precondition, there will be some problems with the context: the context given to first Pre.Prepare only covers its own time limit but not all testcases with it. With the behavior of ssh.Cmd "The command is aborted when ctx's deadline is reached.", the command cannot live across stages (Pre.Prepare, test, Pre.Close). Then one will have to start/stop the process in each stage which is kind of hard to use.

The suggested solution might need more discussion so I have it in a separate cl: crrev.com/c/2174241.

Maybe we can use sed for now to reduce the size of logs? WDYT?

Patch-set: 6