tast-tests/webrtc: don't pass testing.State to peerconnection/ functions
peerconnection/* files get testing.State as parametera -- and also
call s.Fatal() for error situations. It's been mentioned in code
reviews that testing.State is a large struct that should not be
passed around if possible [1], so this CL removes it, relying on
returning error instead.

No functional changes intended.

[1] https://chromium.googlesource.com/chromiumos/platform/tast/+/HEAD/docs/code_review_comments.md#State

BUG=chromium:1071841
TEST=tast run DUT webrtc.RTCPeerConnection.* webrtc.RTCPeerConnectionPerf.*
  on kohaku

Change-Id: I123b4f14f5a4519d3966a0131bb2503c267951ca