tast-tests: predictable order in ReportDiskUsage
ReportDiskUsage logs the largest files present in the directories that
we're specifically watching (/usr/bin, /usr/sbin, ...) to make it easy
to see the ones that have grown a lot between two revisions.
To make it easier to compare to different runs, we iterate over the
directories in the same order every time.

BUG=chromium:1077125
TEST=fast_build.sh -T
TEST=Run "tast run ${DUT_IP} platform.ReportDiskUsage" multiple times,
the output is identical (save for the timestamps)

Change-Id: If2e52b88aae73dd5ed5948b8b91661f1e99cc1e6