Replay functionality for uhid library
Added to the uhid library the possibility of replaying uhid recordings
created using the hid-tools library:
https://gitlab.freedesktop.org/libevdev/hid-tools. There are 2 new
exported functions in this CL NewDeviceFromRecording for creating a new
device from the data contained in a recording and Replay for replaying
events in a recording for an already created device (presumably created
using NewDeviceFromRecording).

BUG=None

TEST=Run the test found in this CL for an example of the working
library.
https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2173302

Change-Id: I77f2a7166cc40b188b91f83854f753be9daeef78