Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The test code shouldn't assume the default window status basically. If another test (with the same precondition) creates a window and turn it into normal status, the next created window should appear as normal, not maximized.

Agree that test code shouldn't assume the state as tast reuse the chrome instance.
Don't we have some utility for a test to do this?

> 
> If LauncherAnimationPerf should perform upon maximized windows, it is LauncherAnimationPerf's responsibility to ensure they are maximized.  Window creation or some sort of things shouldn't care about that.

I also think this should be fixed not in the test code, but in the ash or the code that launch lacros.
Both should behave in the same way even in test?

Patch-set: 2
Reviewer: Gerrit User 1001380 <1001380@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>