tast: Temporarily disable data_collector test
Temporarily disable data_collector.
Build has been red for a while. This test is failing.
Probably this test might not even block the builder since it is not in PFQ.
But the logs indicates some red issues while collecting data and this
test is failing.

Not 100% sure whether it is related, but worth the chance taking into
account that the last green ARC build was 17 days ago.

BUG=chromium:1077329
TEST=tast run arc.DataCollector

Change-Id: I3ec761c1e2c75e3d9c0d8a7e129d2c6adf006b76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2176180
Reviewed-by: Josh Horwich <jhorwich@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: Ricardo Quesada <ricardoq@chromium.org>