Fix flake in ui.ChromeCrashLoggedIn.browser_crashpad_mock_consent
I have seen several flakes like
https://stainless.corp.google.com/browse/chromeos-autotest-results/swarming-4be8779c5cfc9e10/autoserv_test/tast/results/tests/ui.ChromeCrashLoggedIn.browser_crashpad_mock_consent/,
in which the test fails because there is a dmp file and a chrome.txt.gz
file, but no .meta file.

I suggest this is a race condition where the test waits for the root
(browser) Chrome process to exit and then looks for the meta file.
However, crashpad_handler can still be running when the root Chrome
process exits, and crash_reporter can still be running when
crashpad_handler exits. My guess is that the files are still in the
process of being written when the test looks for them.

Wait for crashpad_handler and crash_reporter to finish before looking
for the files.

BUG=None
TEST=tast run ... ui.ChromeCrash*

Change-Id: I538209d7c9f6f59e5de0d913bfa04d0843941494
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2174974
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>