Update patch set 2
Patch Set 2:

This seems generally fine, but I believe that WaitForCrashFiles should be able to handle this race already; it only returns a dict (and a nil err) if all files were created within its deadline, and returns an err if any files were not created within its timeout.

Is this test using WaitForCrashFiles already? If not, is there any particular reason? If so, perhaps a simpler thing to try would be upping the timeout we give it?

Patch-set: 2
Reviewer: Gerrit User 1305461 <1305461@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>