Update patch set 4
Patch Set 4:

(2 comments)

> what's the difference between the 'perf' mechanism here and the timing.json mechanism in Dean's CL?

My understanding:
timing collects the time used in each step of the test (steps can be nested), so it's for debugging the performance of test runtime. (and there's not yet a dashboard for that, I think)
perf package collects metrics and generate results for performance dashboards like crosbolt/chromeperf and can contains metrics other than time (e.g. throughput, ops, etc), so it is the corresponding component of output_perf_value in Autotest.

> are you going to have to update Autotest tko/perf_upload/perf_dashboard_config.json?

I guess it is already included by the tast wildcard in [1]. If not, I'll update it correspondingly.

[1] http://cs/chromeos_public/src/third_party/autotest/files/tko/perf_upload/perf_dashboard_config.json?l=547&rcl=8fd70362934a13f49bff24e0a36af55d99a764cc

Patch-set: 4