Update patch set 12
Patch Set 12:

Sorry for the late reply, but finally found time to add a method to remove users based on hash:

https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2179908

I am still not convinced this is the way to go. We still need to keep the repeated cleanup calls as the owner can not be removed. go/tast-pre-v2 provides a better long-term solution as we can simply remove all users on the device and clear ownership once it lands.

Could we merge this as is to improve the current situation and wait until go/tast-pre-v2 lands?

Patch-set: 12