tast: Let arc.MotionInput expect more than one event at a time
Previously, in the arc.MotionInput test, only the last MotionEvent
received by the test application was reported. We want the ability to
verify multiple events received in a row, because some things (like
pressing the mouse button) are translated into multiple MotionEvents
(ACTION_DOWN and ACTION_BUTTON_PRESS).

We do this by letting the test app report a list of all MotionEvents
received by the app, and adding functionality to clear the reported
events by sending the appropriate intent to the app.

BUG=b:152436345
TEST=tast run DUT arc.MotionInput

Change-Id: I66db8c140d704f84c6f8ca8984582f2b303713d0