Update patch set 1
Patch Set 1:

Thanks Kuang-che, agreed that the block size seems like the metric we care about, even though the difference is likely not that important. As you point out in b/153126253 it's unlikely that the '-B' vs '-b' difference is the only reason we have significantly different sizes between release image and postsubmit images.

More likely different build options, which might be WAI. Maybe that can be revisited?

Patch-set: 1