DualShock 4 test
The test is almost identical to the dualshock 3 test. The code of the
previous commit has been changed to be able to reuse code for both
tests.

The gamepad.DS4 test checks that the mappings on the dualshock 4 are
correct. The uhid library is used to create a virtual dualShock 4,
replay a hid recording and record the received events via a javascript
script. These are then compared with an array of expected button
presses.

BUG=chromium:1081851

TEST=tast -verbose run <DUT_IP> ds3.DS4

Change-Id: Ie0b0400831c595d0b358d008251a1d2714be6f3f