network.HostapHwsim: modify test selections for "sanity" variant
For crbug.com/1049782, we'd like to run the wpa_supplicant "unit tests",
which are actually built into the wpa_supplicant executable and run as
part of the hwsim-based test suite.

For crbug.com/1060000, I've found that one of the 'scan' tests
('scan_only') is flaky. We don't need the full module to run in mainline
-- we only want to make sure the basic integration doesn't regress, and
we can run others as needed.

Trim this to just 3 tests.

BUG=chromium:1049782, chromium:1060000
TEST=`for _ in $(seq 1 100); do
      tast -verbose run -failfortests localhost:9222 \
        network.HostapHwsim.sanity || break
      done`

Change-Id: I6e3afa3d3595a2168d3b765d9fc073751483615f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2183412
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>