tast-tests/webrtc: don't pass testing.State to a mediarecorder/ function
mediarecorder.go gets testing.State as a parameter -- and also
calls s.Fatal() for error situations. It's been mentioned in code
reviews that testing.State is a large struct that should not be
passed around if possible [1], so this CL removes it, relying on
returning error instead.

No functional changes intended.

[1] https://chromium.googlesource.com/chromiumos/platform/tast/+/HEAD/docs/code_review_comments.md#State

BUG=chromium:1071841
TEST=tast run DUT webrtc.MediaRecorderAccelerator.* on kohaku

Change-Id: Idd1fa76efd90b9145aeed443dbdc4e1ae86e3e29